Sync temp thresholds in a stupid way
